requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-jose==3.3.0
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1 